requirements.txt
flask==2.0.1
pandas==1.3.3
mysql-connector-python==8.0.26
python-dotenv==0.19.0
chart.js==4.4.8
requests==2.26.0